requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
celery
redis
boto3
django-storages
python-decouple
pillow
django-allauth
requests
cryptography
google-auth